requirements.txt
markdownify>=0.13.1
beautifulsoup4>=4.12
requests>=2.31
PyPDF2>=3.0
pdfplumber>=0.11
anthropic>=0.40
strands-agents>=0.3
strands-agents-tools>=0.1
python-frontmatter>=1.1
rich>=13.7
click>=8.1
flask>=3.0
whoosh>=2.7
matplotlib>=3.8
